+++ backend/canisters/identity/impl/src/lib.rs
@@ -42,11 +42,6 @@ impl RuntimeState {
         self.data.user_index_canister_id == caller
     }
 
-    pub fn get_principal_from_index(&self, index: u32) -> Principal {
-        let seed = self.data.calculate_seed(index);
-        self.get_principal_from_seed(seed)
-    }
-
     pub fn get_principal_from_seed(&self, seed: [u8; 32]) -> Principal {
         let public_key = self.der_encode_canister_sig_key(seed);
         Principal::self_authenticating(public_key)
@@ -55,6 +50,17 @@ impl RuntimeState {
     pub fn der_encode_canister_sig_key(&self, seed: [u8; 32]) -> Vec<u8> {
         let canister_id = self.env.canister_id();
         CanisterSigPublicKey::new(canister_id, seed.to_vec()).to_der()
+    }
+
+    pub fn push_new_user(&mut self, auth_principal: Principal, originating_canister: CanisterId) -> (Principal, [u8; 32]) {
+        let index = self.data.user_principals.next_index();
+        let seed = self.data.calculate_seed(index);
+        let principal = self.get_principal_from_seed(seed);
+        self.data
+            .user_principals
+            .push(index, principal, auth_principal, originating_canister);
+
+        (principal, seed)
     }
 
     pub fn metrics(&self) -> Metrics {
@@ -78,22 +84,15 @@ struct Data {
     governance_principals: HashSet<Principal>,
     user_index_canister_id: CanisterId,
     cycles_dispenser_canister_id: CanisterId,
-    #[serde(default = "internet_identity_canister_id")]
     internet_identity_canister_id: CanisterId,
     user_principals: UserPrincipals,
     legacy_principals: HashSet<Principal>,
     #[serde(skip)]
     signature_map: SignatureMap,
-    #[serde(default)]
     salt: Salt,
-    #[serde(default)]
     principal_migration_job_enabled: bool,
     rng_seed: [u8; 32],
     test_mode: bool,
-}
-
-fn internet_identity_canister_id() -> CanisterId {
-    CanisterId::from_text("rdmx6-jaaaa-aaaaa-aaadq-cai").unwrap()
 }
 
 impl Data {